requirements.txt
tensorflow==2.18.0
keras==3.8.0
streamlit
opencv-python-headless
matplotlib
pillow
reportlab
google-api-python-client
google-auth-oauthlib
python-dotenv
pandas
numpy
supabase
streamlit-authenticator
